+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This is the official BubbleOS `CHANGELOG`! All features will be recorded in _'gr
 
 - Changed the confirmation prompt to automatically enter when you press a key.
 - Fixed an issue where BubbleOS would crash when using `mkfile`/`mkdir` if the path exceeded `MAX_PATH`.
+- Fixed an issue where making a directory/file with invalid characters would crash BubbleOS.
 - BubbleOS will now crash with a fatal error if the timebomb has expired, after displaying a message.
 
 ### Removed Features